reorganise flexiwidget, non-working
