+++ packages/svelte-flexiboards/src/lib/system/announcer.svelte.ts
@@ -1,5 +1,5 @@
 import type { InternalFlexiBoardController } from './provider.svelte.js';
-import { generateUniqueId } from './utils.svelte.js';
+import { generateUniqueId } from './shared/utils.svelte.js';
 
 export type AriaPoliteness = 'off' | 'polite' | 'assertive';
